upgrade with more changes from luberjack v2

+++ AUTHOR_SETUP.js
@@ -1,7 +1,11 @@
 const fs = require('fs');
 const path = require('path');
+const { execSync } = require('child_process');
 
 const fgGreen = '\x1b[32m';
+const fgYellow = '\x1b[33m';
+const fgWhite = '\x1b[37m';
+
 const driverName = process.argv.slice(2)[0] || 'lumberjack-custom-driver';
 
 const filesToIgnore = {
@@ -35,6 +39,12 @@ const cammelName = splittedName
 const uppercaseUnderscoreName = splittedName.map((name) => name.toUpperCase()).join('_');
 
 function swapFileNames(dir) {
+  _swapFileNames(dir);
+
+  console.info(fgWhite, 'Path driver names swap... Completed');
+  console.info(fgWhite, 'Content driver names swap... Completed');
+}
+function _swapFileNames(dir) {
   try {
     fs.readdirSync(dir).forEach(function (file) {
       let filePath = path.normalize(path.join(dir, file));
@@ -42,7 +52,7 @@ function swapFileNames(dir) {
       if (!filesToIgnore[file]) {
         const isDir = fs.lstatSync(filePath).isDirectory();
         if (isDir) {
-          swapFileNames(filePath);
+          _swapFileNames(filePath);
         } else {
           replaceNameOccurrences(filePath);
         }
@@ -82,6 +92,8 @@ function replaceNameOccurrences(filePath) {
 function renameFilenamePlaceholder(filePath) {
   if (filePath.includes(hyphenFileToken)) {
     const newPath = filePath.replace(new RegExp(hyphenFileToken, 'g'), hyphenName);
+
+    fs.rmdirSync(newPath, { recursive: true });
     fs.renameSync(filePath, newPath);
     return newPath;
   }
@@ -107,14 +119,39 @@ function updatePackageJson() {
     };
 
     fs.writeFileSync('package.json', JSON.stringify(packageJson, null, 2));
+
+    console.info(fgWhite, 'package.json update... Completed');
   } catch (error) {
     console.log(error);
   }
 }
 
 function activateGithubActions() {
-  fs.renameSync('github', '.github');
+  if (fs.existsSync('github')) {
+    fs.rmdirSync('.github', { recursive: true });
+    fs.renameSync('github', '.github');
+  }
+
+  console.info(fgWhite, 'Github Actions activation... Completed');
 }
+
+function fixLintIssues() {
+  execSync('yarn lint --fix');
+
+  console.info(fgWhite, 'Fix linting issue... Completed');
+}
+
+function renameRootFolder() {
+  const splittedPath = __dirname.split('/');
+  splittedPath[splittedPath.length - 1] = hyphenName;
+  const newPath = path.normalize(splittedPath.join('/'));
+
+  fs.renameSync(__dirname, newPath);
+
+  console.info(fgWhite, 'Rename root folder... Completed\n');
+}
+
+console.info(fgYellow, '\nPlease wait. This will take a minute ⏳\n');
 
 swapFileNames(__dirname);
 
@@ -122,4 +159,8 @@ updatePackageJson();
 
 activateGithubActions();
 
-console.info(fgGreen, '✅ Success: You can now implement your driver 🚀');
+fixLintIssues();
+
+renameRootFolder();
+
+console.info(fgGreen, '✅ Success: You can now implement your driver 🚀\n');
